requirements.txt
py-cord==2.3.2
aiohttp==3.8.6
python-dotenv==1.0.1
Flask==3.0.3
discord.py==2.6.2
PyGithub

